fix: update entity mappings for database schema

+++ src/main/java/com/example/demo/song/SongRepository.java
@@ -2,5 +2,5 @@ package com.example.demo.song;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SongRepository extends JpaRepository<Song, Integer> {
+public interface SongRepository extends JpaRepository<Song, Long> {
 }
